Slight change in lists affecting indented code in lists.
We now parse an optional space after parsing the initial
space indent and the width of the marker.  This means that, if
you start your main text one space after the list marker,
an indented code block will need to be indented four spaces
beyond that, not four spaces beyond the space between the
list marker and the text.  That is more intuitive, I think.