Removed "two spaces break out" rule and tests.
I'm no longer convinced this is needed, and it raises complexity
of parser considerably, since we have to consider also cases where
the lines are not really blank but have blockquote starts, etc.

We split blockquotes if they have a blank between.  As for lists,
I think that the two-blank rule is perhaps no longer needed,
since we make a new list when the bullet or number style
changes, and when you move from tight to list.

In any case it can be implemented in listParser if desired.